// packages/core/src/types/floating-ui.ts
// Placement copied from floating-ui to avoid template literal type
// which is unsupported in ts v3
export type Placement =
  | 'top'
  | 'right'
  | 'bottom'
  | 'left'
  | 'top-start'
  | 'top-end'
  | 'right-start'
  | 'right-end'
  | 'bottom-start'
  | 'bottom-end'
  | 'left-start'
  | 'left-end';